[Bronze V] Title: 이상한 기호, Time: 100 ms, Memory: 14136 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -8,7 +8,7 @@ public class Main {
             StringTokenizer st = new StringTokenizer(br.readLine());
             int A = Integer.parseInt(st.nextToken());
             int B = Integer.parseInt(st.nextToken());
-            int answer = (A+B)*(A-B);
+            long answer = (long) (A+B)*(A-B);
             
             bw.write(String.valueOf(answer));
             bw.flush();
